Tests mad on configure_register function and read_register checking

+++ main/mlx90640.h
@@ -10,7 +10,7 @@
 
 // Functions prototypes
 bool mlx_initial_setup(i2c_buffer_type *);
-bool mlx_transmit_receive(i2c_buffer_type *, uint8_t, uint8_t);
+bool mlx_transmit_receive(i2c_buffer_type *, uint8_t, uint16_t);
 bool mlx_transmit(i2c_buffer_type *, uint8_t);
 bool mlx_configure_register(i2c_buffer_type *, uint16_t, uint16_t, bool);
 bool mlx_read_register(i2c_buffer_type *, uint16_t, uint16_t *);
